# optimized bbox_to_mask3d by vectorizing loop by 5526.89%