completed todo by vectorizing the main loop in bbox_to_mask3d(), leading to speed improvements of up to 5526.89 percent on average on CUDA

Benchmarking on device: CPU, B=128, shape=(32, 32, 32)
Original Time: 0.116394 s
Vectorized Time: 0.058965 s
percentage improvement = **197.39%**

Benchmarking on device: CUDA, B=128, shape=(32, 32, 32)
Original Time: 0.090199 s
Vectorized Time: 0.001632 s
percentage speedup = **5526.89%**

https://colab.research.google.com/drive/1BgU0yma9UkThbegWORgId-CM_uGSu0aQ?usp=sharing

Here's a Colab notebook with the original and the vectorized version. This also checks and proves that both versions are identical in terms of output

- [x] optimization of previously written code

#### Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] Did you update CHANGELOG in case of a major change?