@sidd-27 2025-08-05
I did register for it, but not really, I just wanted to get into open source, and a friend said this was a good repo, so I started doing some of the todos left in the code, working on a quartic solver rn. (but if I'm winning something then yes lol 😂)